requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
mysqlclient>=2.2
python-decouple>=3.8
django-cors-headers>=4.3
openai>=1.84.0
Pillow>=11.2.1